add second line to text
this required a change in the way the text object was
positioned. before, we had the position outside the
text object, now it is inside using the text matrix
command Tm.

we also have to set the text leading by using the
TL command. Without this, the next line operator
T* would only bring the cursor back but not down.