fiddle with text matric and leading parameter to get everything on the page
the text matrix governs the positioning of text on the page. its e and f
parameters could be interpreted as x and y coordinates with the origin
located at the south west position of a page, I think.

I set it to e and f to 30 and 750 now, starting text rendering very
close to the top-left, north-west border.

the leading parameter governs the line distance. its value is somewhat
dependent on the font size. we have 12, so I set the leading to 16,
which gives us a line distance which looks ok-ish.