main, pdf: flush also on V grout command
this uncovers an interessting bug, where it gets apparent, that we will
not get around implementing a correct relative positioning:

the second '-' character of the page number indication at the bottom of
the page is rendered exactly above the first and leading one.

this is because, there is no exact positioning, but an interspersed V,
where we now also put an Tm command.