Transpiler, pdf: fix bug where fonts were added doubly in Document
we use the Transpiler font map to remember which fonts were already added.
grout actually helps us in issuing `x font N FF` commands on each new
page it seems. the `N` stays stable across pages.

but the bug is that we add that font also to the global font list again,
which is not necessary and leads to wrong font selection and rendering
in pdf viewers.

another bug not fixed by this change is that the logical pdf font number
changes on the second page for `CR` from `3` on page 1 to `1` on page 2.
this is not reflected in the change command.