Transpiler,groff,pdf: add basic drawing support - `Dl` grout
add just enough drawing support to be able to draw the page header lines
at the top of the page.

introduce a new struct `GraphicalObject`, which function much in the
same way as a `TextObject`, but obviously with functions for drawing,
currently only `lineTo`.

`GraphicalObject`s are part of `Stream` objects and are rendered in the
final output _before_ the `TextObject`.